# services/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kobe API",
    "version": "0.0.0",
    "description": "Team-scoped routes act on the session's active team (spec D9); send X-Kobe-Team on changes. Errors are `{code, message}`. Authentication is the Better Auth session cookie."
  },
  "paths": {
    "/v1/threads": {
      "get": {
        "operationId": "listThreads",
        "summary": "List the viewer's threads (or a project's readable threads), newest first",
        "description": "Keyset pagination on (last_activity_at, id). Never Trash. With `q`, full-text search (KOBE-33): web-search syntax (`\"phrase\"`, `-exclude`, `or`) over user/assistant messages (all terms in one message) and titles (plus fuzzy title matching), ranked by relevance with snippets; `cursor` is then a search cursor and `limit` is capped at 50. Reads Postgres only (never wakes a sandbox).",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,3}$"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": false,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of threads (`ThreadSearchPage` when `q` is given).",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ThreadPage"
                    },
                    {
                      "$ref": "#/components/schemas/ThreadSearchPage"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_cursor`, `invalid_query` (only excluded terms) or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role) or `read_only`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`no_active_team` or `team_mismatch` (stale tab), or a thread conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "`search_timeout`: the search exceeded its time limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createThread",
        "summary": "Create a thread in the active team, owned by the caller",
        "parameters": [
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_cursor` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role) or `read_only`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`agent_not_found` or `project_not_found`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`no_active_team` or `team_mismatch` (stale tab), or a thread conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/threads/{id}": {
      "get": {
        "operationId": "getThread",
        "summary": "A thread with its entry tree (append order), leaf, agent version and status",
        "description": "Readable by its owner, or by project members when shared. Never wakes a sandbox.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]{0,9})$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,3}$"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": false,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The thread and the first page of entries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDetail"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_cursor` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role) or `read_only`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found`: unknown id, another team's or user's thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`no_active_team` or `team_mismatch` (stale tab), or a thread conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateThread",
        "summary": "Rename, or share to / unshare from its project (owner only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThreadBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_cursor` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role) or `read_only`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found`: unknown id, another team's or user's thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`no_active_team` or `team_mismatch` (stale tab), or a thread conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "trashThread",
        "summary": "Move to Trash (30-day soft delete; owner only; idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The trashed thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_cursor` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role) or `read_only`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found`: unknown id, another team's or user's thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`no_active_team` or `team_mismatch` (stale tab), or a thread conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/threads/{id}/messages": {
      "post": {
        "operationId": "submitMessage",
        "summary": "Send a message: starts a run, or queues it behind the thread's active run (D17)",
        "description": "Owner only. The run is created queued and starts at once when nothing is active and the thread is not interrupted; otherwise `queued` is true and it starts in order (Retry runs first). Follow it with `GET /v1/runs/{id}/events`. `parent_entry_id` branches from an entry (edit-and-regenerate); absent = the thread's leaf when the run starts.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitMessageBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The run, and whether it waits in the queue.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitMessageResult"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role, or a deactivated account), or `read_only` (a project reader of a shared thread).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found` or `entry_not_found`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`thread_in_trash`, `queue_full`, `no_active_team`, `team_mismatch`, or `thread_busy`: try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`attachments_unavailable`: file_ids before uploads exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "`budget_exhausted`: no new runs at 100 % of a budget (D30).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "`isolation_unavailable`: agents are disabled (D4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/threads/{id}/runs": {
      "get": {
        "operationId": "listThreadRuns",
        "summary": "The thread's active run and its queued runs, in start order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": false,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active run first, then the queue (`queue_pos` 1 = next).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadRuns"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role, or a deactivated account), or `read_only` (a project reader of a shared thread).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/threads/{id}/queue/resume": {
      "post": {
        "operationId": "resumeThreadQueue",
        "summary": "Continue without retry: an interrupted thread's queued runs resume (D14)",
        "description": "No-op unless the thread is interrupted. Returns the thread's runs.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The thread's runs after resuming.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadRuns"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role, or a deactivated account), or `read_only` (a project reader of a shared thread).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`thread_not_found`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`thread_in_trash` or `thread_busy`: try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/runs/{id}/steer": {
      "post": {
        "operationId": "steerRun",
        "summary": "Steer now: inject a message into the active run at Pi's next safe point (D17)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Kobe-Team",
            "in": "header",
            "required": true,
            "description": "The team the client believes is active. Required on changes; must match.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SteerBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The run; `steer.applied` is on its event stream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` or `team_header_required`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated`: no session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`not_a_team_member`, `forbidden` (team role, or a deactivated account), or `read_only` (a project reader of a shared thread).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "`run_not_found`: unknown id, or a run of another team's or another user's thread.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "`invalid_transition` (not running in the workspace) or `thread_busy`: try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "`sandbox_unavailable`: the workspace did not answer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "additionalProperties": false
      },
...
      "CreateThreadBody": {
        "type": "object",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "pattern": "^[^\\u0000-\\u001f\\u007f]*$"
          }
        },
        "additionalProperties": false
      },
      "UpdateThreadBody": {
        "type": "object",
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 200,
                "pattern": "^[^\\u0000-\\u001f\\u007f]*$"
              },
              {
                "type": "null"
              }
            ]
          },
          "shared_to_project": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SubmitMessageResult": {
        "type": "object",
        "properties": {
          "run_id": {
            "type": "string",
            "allOf": [
              {
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "pattern": "^[0-9a-f-]+$"
              }
            ]
          },
          "queued": {
            "type": "boolean"
          }
        },
        "required": ["run_id", "queued"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}